Пропустил поля которые нужно изменить в файле миграции

+++ src/main/resources/db/migration/V1__Init_DB.sql
@@ -5,12 +5,12 @@ create table message (
     filename varchar(255),
     tag varchar(255),
     text varchar(2048) not null,
-    user_id int8,
+    user_id bigint,
     primary key (id)
 );
 
 create table user_role (
-    user_id int8 not null,
+    user_id bigint not null,
     roles varchar(255)
 );
